+++ src/sass/main.css
@@ -1,0 +1,4 @@
+body {
+  background-color: #8f8c8c;
+}
+/*# sourceMappingURL=main.css.map */
