+++ src/sass/main.scss
@@ -1,0 +1,3 @@
+ body {
+     background-color: #8f8c8c;
+ }
